refactor: move MentionSuggestions files to its own directory
