+++ src/script/page/message-list/MentionSuggestions/MentionSuggestions.tsx
@@ -24,9 +24,9 @@ import {KEY} from 'Util/KeyboardUtil';
 import {clamp} from 'Util/NumberUtil';
 import useEffectRef from 'Util/useEffectRef';
 
-import {useFadingScrollbar} from '../../ui/fadingScrollbar';
-import MentionSuggestionsItem from './MentionSuggestions/MentionSuggestionsItem';
-import {User} from '../../entity/User';
+import {useFadingScrollbar} from '../../../ui/fadingScrollbar';
+import MentionSuggestionsItem from './MentionSuggestionsItem';
+import {User} from '../../../entity/User';
 
 type MentionSuggestionListProps = {
   onSelectionValidated: (data: User, element: HTMLInputElement) => void;
